Fix min_length/max_length validation of string fields
In the case of a string that was not a representation of an integer, the
validation of min_length and max_length was entirely ignored.
e.g. 'x01234567890123456' passes max_length = 3 validation, because it
cannot be converted to int

If the string was an integer representation, then it was parsed to
integer and converted to unicode again for length comparison. This
resulted in the stripping of all leading zeroes.
e.g. the value '0123' failed min_length = 4 validation, because '123' is
only three chars long

The fix moves the conversion to an intermediate subclass, which
MinValueValidator and MaxValueValidator explicitly derive from, because
they require this behaviour.